requirements.txt
playwright
yt-dlp
openai-whisper
